remove get exec times and timetable store functins

+++ exec_tracker.py
@@ -8,46 +8,47 @@ would run it afterwards at 6AM, 7AM etc...
 
 from datetime import datetime
 import json
+from os import path
 
 DATE_FORMAT = '%m/%d/%Y, %H:%M:%S'
 TIMETABLE_PATH = 'EXEC_TIMETABLE.json'
 
+
 class ExecTracker:
     def __init__(self):
-        try:
+        self.timetable_modified = False
+
+        if path.exists(TIMETABLE_PATH):
             self.timetable = open(TIMETABLE_PATH, 'r+')
-        except FileNotFoundError:
+            self.exec_times = json.load(self.timetable)
+        else:
             self.timetable = open(TIMETABLE_PATH, 'w')
             self.timetable.write('{}')
             self.timetable.close()
-            self.timetable = open(TIMETABLE_PATH, 'r+')
-        finally:
-            self.exec_times = json.load(self.timetable)
+            self.exec_times = {}
 
     def __del__(self):
-        self.timetable.close()
+        if self.timetable_modified:
+            self.timetable = open(TIMETABLE_PATH, 'r+')
+            json.dump(self.exec_times, self.timetable, indent=4)
+            self.timetable.close()
 
-    def get_exec_times(self):
-        return self.exec_times
-
-    def script_executed_today(self, exec_times, location):
-        if location.name in exec_times:
-            for exec_time_str in exec_times[location.name]:
+    def script_executed_today(self, location):
+        if location.name in self.exec_times:
+            for exec_time_str in self.exec_times[location.name]:
                 exec_time = datetime.strptime(exec_time_str, DATE_FORMAT)
                 if exec_time.date() == location.get_local_time_utc().date():
                     return True
 
         return False
 
-    def append_exec_time(self, exec_times, location):
+    def mark_exec_time(self, location):
+        self.timetable_modified = True
+
         location_name = location.name
         local_time_utc_str = location.get_local_time_utc().strftime(DATE_FORMAT)
 
-        if location_name in exec_times:
-            exec_times[location_name].append(local_time_utc_str)
+        if location_name in self.exec_times:
+            self.exec_times[location_name].append(local_time_utc_str)
         else:
-            exec_times[location_name] = [local_time_utc_str]
-
-    def timetable_store(self, exec_times):
-        self.timetable.seek(0)
-        json.dump(exec_times, self.timetable, indent=4)
+            self.exec_times[location_name] = [local_time_utc_str]
